Refactor tests and minor improvements
